Worked on it again
Ingredientquantity.quantity needs to be a double, not a float, so that I
can accurately add and subtract.
Right now it crashes but I think it somewhat adds ingredient quantities,
but one of the additions is crashing it.

+++ app/src/main/java/edu/calvin/cs262/scrumptious/IngredientQuantity.java
@@ -12,7 +12,7 @@ public class IngredientQuantity {
     // Data
     private Ingredient ingredient;
     private String unit;
-    private float quantity;
+    private double quantity;
 
     // Constructor
     public IngredientQuantity(Ingredient ingredient, String unit, float quantity) {
@@ -38,11 +38,11 @@ public class IngredientQuantity {
         this.unit = unit;
     }
 
-    public float getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(float quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 }
